--- classes/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'jamesma'
-__author__ = 'chicago-joe'
